requirements.txt
pytest==7.1.2
selenium==4.1.3
pytest-check==1.0.9
allure-pytest==2.9.45